+++ src/main/java/com/runninghi/feedback/command/domain/service/FeedbackService.java
@@ -1,7 +1,5 @@
 package com.runninghi.feedback.command.domain.service;
 
-import com.runninghi.feedback.command.application.dto.FeedbackNoDTO;
-import com.runninghi.feedback.command.application.dto.FeedbackReplyDTO;
 import com.runninghi.feedback.command.application.dto.SaveFeedbackDTO;
 import jakarta.transaction.Transactional;
 
@@ -11,14 +9,4 @@ public interface FeedbackService {
     @Transactional
     Long saveFeedback(SaveFeedbackDTO feedbackDTO, Long userNo);
 
-    @Transactional
-    Long saveFeedbackReply(FeedbackReplyDTO feedbackReplyDTO);
-
-    // 피드백 답변 삭제
-    @Transactional
-    Long deleteFeedbackReply(FeedbackNoDTO feedbackNoDTO);
-
-    // 피드백 답변 수정
-    @Transactional
-    Long modifyFeedbackReply(FeedbackReplyDTO feedbackReplyDTO);
 }
